Fixes syntax for checking if pre-commit is installed in isaaclab.sh (#1422)
# Description

When running `./isaaclab.sh --format` for another pull request, I got a
`command not found` error for `pre-commit`. Looking into the script, I
found a condition to check if `pre-commit` was installed, but the syntax
was off.

It was written like this:

```bash
if [ ! command -v pre-commit &>/dev/null ]; then
```

But it should be:

```bash
if ! command -v pre-commit &>/dev/null; then
```

I fixed it, and the script worked as expected, installing `pre-commit`
when I ran it again.

## Type of change

- Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./isaaclab.sh --format`
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there

---------

Signed-off-by: Kelly Guo <kellyg@nvidia.com>
Signed-off-by: Kelly Guo <kellyguo123@hotmail.com>
Co-authored-by: Kelly Guo <kellyg@nvidia.com>
Co-authored-by: Kelly Guo <kellyguo123@hotmail.com>