Updates code snippets in documentation to reference scripts (#1693)
# Description

This change replaces some of the hardcoded python snippets in the sensor
documentation to reference directly from the demo scripts. This avoids
inconsistencies when changes in the scripts or documentations are made.


## Type of change

- This change requires a documentation update

## Checklist

- [X] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./isaaclab.sh --format`
- [X] I have made corresponding changes to the documentation
- [X] My changes generate no new warnings
- [X] I have added tests that prove my fix is effective or that my
feature works
- [X] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [X] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there

<!--
As you go through the checklist above, you can mark something as done by
putting an x character in it

For example,
- [x] I have done this task
- [ ] I have not done this task
-->