Clarifies layer instructions in animation docs (#2240)
# Description

I tried following the "Recording Animations of Simulations" instructions
(https://isaac-sim.github.io/IsaacLab/main/source/how-to/record_animation.html)

Most of the instructions were very clear. I was able to create the
`Stage.usd` and `TimeSample_tk001.usd` files. However, I struggled on
this step:

"On a new stage, add the Stage.usd as a sublayer and then add the
TimeSample_tk001.usd as a sublayer. You can do this by dragging and
dropping the files from the file explorer to the stage. "

I submitted an issue and it was clear that a documentation update would
be helpful.

Fixes #2232

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./isaaclab.sh --format`
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] (Docs only) I have added tests that prove my fix is effective or
that my feature works
- [x] (Docs only) I have updated the changelog and the corresponding
version in the extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there

Co-authored-by: Mayank Mittal <12863862+Mayankm96@users.noreply.github.com>